experiments/internet_retrive/tavily_exp/pyproject.toml
[project]
name = "tavily-exp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
